renamed some unittest as described in the framework_unittest/README
