+++ .travis.yml
@@ -28,9 +28,9 @@ before_script:
 
 #run tests
 script:
-  - python -m pytest --cov-append --cov=./ framework_unittests/test_Server_2.py
-  - python -m pytest --cov-append --cov=./ framework_unittests/test_Server.py
+  - python -m pytest --cov-append --cov=./ framework_unittests/test_A_Server_2.py
+  - python -m pytest --cov-append --cov=./ framework_unittests/test_A_Server.py
   - python -m pytest --cov-append --cov=./ test_server_alive.py
-  - python -m pytest --ff -n 2 --cov-append --cov=./  framework_unittests/test_IPC.py framework_unittests/test_Router.py framework_unittests/test_Yaml.py framework_unittests/test_Logger.py framework_unittests/test_firmware_handler.py
+  - python -m pytest --ff -n 2 --cov-append --cov=./  framework_unittests/test_A_IPC.py framework_unittests/test_AP_Router.py framework_unittests/test_AP_Yaml.py framework_unittests/test_AP_Logger.py framework_unittests/test_AP_firmware_handler.py
 
   - "python -m pep8 --exclude=migrations --ignore=E501,E225 */*.py *.py"
