+++ framework_unittests/test_Server_VLAN.py
@@ -1,5 +1,5 @@
 
-from framework_unittests.test_Server_2 import ServerCore, block_until_server_is_online
+from framework_unittests.test_A_Server_2 import ServerCore, block_until_server_is_online
 from server.server import Server
 from multiprocessing import Process
 import unittest
